backend/requirements.txt
# FastAPI + Server
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
orjson==3.11.1

# Embeddings + Semantic Search
# scikit-learn==1.6.1
# numpy==2.0.2

# Vector DB
pinecone==7.3.0

# LangChain (for optional chunking + loading)
langchain==0.3.27
langchain-text-splitters==0.3.9

# Document parsing
PyMuPDF==1.26.3       # PDFs
python-docx==1.2.0    # Word docs
lxml==6.0.0           # HTML/EML

# Utilities
tqdm==4.67.1
requests==2.32.4
pydantic==2.11.7
typing_extensions==4.14.1